+++ tests/Integration/DSC_SqlAlwaysOnService.Integration.Tests.ps1
@@ -54,7 +54,7 @@ AfterAll {
 
 <#
     TODO: This has temporarily been disabled as the test is not passing.
-          Tags should be changed to @('Integration_SQL2016', 'Integration_SQL2017', 'Integration_SQL2019')
+          Tags should be changed to @('Integration_SQL2016', 'Integration_SQL2017', 'Integration_SQL2019', 'Integration_SQL2022')
 #>
 Describe "$($script:dscResourceName)_Integration" -Tag 'Skip'  {
     BeforeAll {