+++ CHANGELOG.md
@@ -28,6 +28,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - New GitHub issue templates for proposing new public commands, proposing
     an enhancement to an existing command, or having a problem with an existing
     command.
+  - Integration tests are now also run on SQL Server 2022.
+- SqlDatabase
+  - Added compatibility levels for SQL Server 2022 (major version 16).
+- SqlSetup
+  - Paths for SQL Server 2022 are correctly returned by Get.
 
 ### Changed